backend/requirements.txt
aiogram>=3.0.0
langchain
langchain-openai
supabase
pandas
python-dotenv
pdfplumber
